feat: marcar uma resposta como solucao

+++ src/main/java/br/com/forum_hub/domain/resposta/RespostaService.java
@@ -2,16 +2,26 @@ package br.com.forum_hub.domain.resposta;
 
 import br.com.forum_hub.domain.topico.Status;
 import br.com.forum_hub.domain.topico.TopicoService;
+import br.com.forum_hub.domain.usuario.Usuario;
 import br.com.forum_hub.infra.exception.RegraDeNegocioException;
 import jakarta.transaction.Transactional;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
 import org.springframework.stereotype.Service;
 
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 
 @Service
 public class RespostaService {
     private final RespostaRepository repository;
     private final TopicoService topicoService;
+
+    @Autowired
+    private RoleHierarchy roleHierarchy;
+
+    @Autowired
+    private HierarquiaService hierarquiaService;
 
     public RespostaService(RespostaRepository repository, TopicoService topicoService) {
         this.repository = repository;
@@ -22,11 +32,11 @@ public class RespostaService {
     public Resposta cadastrar(DadosCadastroResposta dados, Long idTopico) {
         var topico = topicoService.buscarPeloId(idTopico);
 
-        if(!topico.estaAberto()) {
+        if (!topico.estaAberto()) {
             throw new RegraDeNegocioException("O tópico está fechado! Você não pode adicionar mais respostas.");
         }
 
-        if(topico.getQuantidadeRespostas() == 0) {
+        if (topico.getQuantidadeRespostas() == 0) {
             topico.alterarStatus(Status.RESPONDIDO);
         }
 
@@ -37,20 +47,28 @@ public class RespostaService {
     }
 
     @Transactional
-    public Resposta atualizar(DadosAtualizacaoResposta dados) {
+    public Resposta atualizar(DadosAtualizacaoResposta dados, Usuario logado) throws AccessDeniedException {
         var resposta = buscarPeloId(dados.id());
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, resposta.getTopico().getAutor(), "ROLE_MODERADOR"))
+            throw new AccessDeniedException("Você não pode editar essa resposta!");
+
         return resposta.atualizarInformacoes(dados);
     }
 
-    public List<Resposta> buscarRespostasTopico(Long id){
+    public List<Resposta> buscarRespostasTopico(Long id) {
         return repository.findByTopicoId(id);
     }
 
     @Transactional
-    public Resposta marcarComoSolucao(Long id) {
+    public Resposta marcarComoSolucao(Long id, Usuario logado) throws AccessDeniedException {
         var resposta = buscarPeloId(id);
 
         var topico = resposta.getTopico();
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_INSTRUTOR"))
+            throw new AccessDeniedException("Você não pode marcar essa resposta como solução!");
+
         if(topico.getStatus() == Status.RESOLVIDO)
             throw new RegraDeNegocioException("O tópico já foi solucionado! Você não pode marcar mais de uma resposta como solução.");
 
@@ -58,10 +76,14 @@ public class RespostaService {
         return resposta.marcarComoSolucao();
     }
 
+
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) throws AccessDeniedException {
         var resposta = buscarPeloId(id);
         var topico = resposta.getTopico();
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR"))
+            throw new AccessDeniedException("Você não pode apagar essa resposta!");
 
         repository.deleteById(id);
