+++ src/main/java/br/com/forum_hub/domain/topico/TopicoService.java
@@ -1,19 +1,26 @@
 package br.com.forum_hub.domain.topico;
 
 import br.com.forum_hub.domain.curso.CursoService;
+import br.com.forum_hub.domain.resposta.HierarquiaService;
 import br.com.forum_hub.domain.usuario.Usuario;
 import br.com.forum_hub.infra.exception.RegraDeNegocioException;
 import jakarta.transaction.Transactional;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+
+import java.nio.file.AccessDeniedException;
 
 @Service
 public class TopicoService {
 
     private final TopicoRepository repository;
     private final CursoService cursoService;
+
+    @Autowired
+    private HierarquiaService hierarquiaService;
 
     public TopicoService(TopicoRepository repository, CursoService cursoService) {
         this.repository = repository;
@@ -38,15 +45,23 @@ public class TopicoService {
     }
 
     @Transactional
-    public Topico atualizar(DadosAtualizacaoTopico dados) {
+    public Topico atualizar(DadosAtualizacaoTopico dados, Usuario logado) throws AccessDeniedException {
         var topico = buscarPeloId(dados.id());
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR"))
+            throw new AccessDeniedException("Você não pode editar esse tópico!");
+
         var curso = cursoService.buscarPeloId(dados.cursoId());
         return topico.atualizarInformacoes(dados, curso);
     }
 
     @Transactional
-    public void excluir(Long id) {
+    public void excluir(Long id, Usuario logado) throws AccessDeniedException {
         var topico = buscarPeloId(id);
+
+        if(hierarquiaService.usuarioNaoTemPermissoes(logado, topico.getAutor(), "ROLE_MODERADOR"))
+            throw new AccessDeniedException("Você não pode apagar esse tópico!");
+
         if (topico.getStatus() == Status.NAO_RESPONDIDO)
             repository.deleteById(id);
         else
